requirements/app.txt
annotated-types==0.6.0
anyio==3.7.1
bleach==6.1.0
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.104.1
future==0.18.3
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
pydantic==2.4.2
pydantic_core==2.10.1
python-multipart==0.0.6
python-whois==0.8.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0
webencodings==0.5.1
slowapi==0.1.8
